[tables] Migrate perf tests to service level
