+++ sdk/tables/data-tables-perf-tests/tsconfig.src.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../../tsconfig.lib.json"
+}
