--- sdk/tables/perf-tests/data-tables/tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "extends": "../../../../tsconfig",
-  "compilerOptions": {
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext",
-    "rootDir": ".",
-    "paths": {
-      "@azure-tests/perf-data-tables": ["./src/index"]
-    }
-  },
-  "include": [
-    "src/**/*.ts",
-    "src/**/*.mts",
-    "src/**/*.cts",
-    "samples-dev/**/*.ts",
-    "src/**/*.ts",
-    "src/**/*.mts",
-    "src/**/*.cts"
-  ]
-}
